ai_chat/pyproject.toml
[project]
name = "ai-chat-cultural-assistant"
version = "0.1.0"
description = "AI cultural chat system with image recognition, ethnic clothing analysis, and minority language translation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
